pcb_io: fix clang warnings across all import/export plugins
Remove unused variables, fix sign-compare mismatches, resolve shadow
declarations, and eliminate unnecessary copies in range-for loops.

+++ pcbnew/pcb_io/easyeda/pcb_io_easyeda_plugin.cpp
@@ -327,7 +327,7 @@ void PCB_IO_EASYEDA::FootprintEnumerate( wxArrayString&  aFootprintNames,
 
                     std::map<wxString, wxString> paramMap;
 
-                    for( int i = 1; i < paramParts.size(); i += 2 )
+                    for( int i = 1; i < (int) paramParts.size(); i += 2 )
                     {
                         wxString key = paramParts[i - 1];
                         wxString value = paramParts[i];
@@ -434,7 +434,7 @@ FOOTPRINT* PCB_IO_EASYEDA::FootprintLoad( const wxString& aLibraryPath,
 
                     std::map<wxString, wxString> paramMap;
 
-                    for( int i = 1; i < paramParts.size(); i += 2 )
+                    for( int i = 1; i < (int) paramParts.size(); i += 2 )
                     {
                         wxString key = paramParts[i - 1];
                         wxString value = paramParts[i];
